pylint fixes and renaming typing.py to models.py to avoid conflicts
