+++ tests/integration/test_integration.py
@@ -32,8 +32,8 @@ import pytest
 
 # First-Party
 from mcpgateway.main import app, require_auth
+from mcpgateway.models import InitializeResult, ResourceContent, ServerCapabilities
 from mcpgateway.schemas import ResourceRead, ServerRead, ToolMetrics, ToolRead
-from mcpgateway.types import InitializeResult, ResourceContent, ServerCapabilities
 
 
 # -----------------------------------------------------------------------------
