--- tests/unit/mcpgateway/db/test_models.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-
-Copyright 2025
-SPDX-License-Identifier: Apache-2.0
-Authors: Mihai Criveti
-
-"""
