+++ tests/unit/mcpgateway/test_models.py
@@ -6,7 +6,7 @@ SPDX-License-Identifier: Apache-2.0
 Authors: Mihai Criveti
 
 This module contains tests for the various MCP protocol type definitions
-defined in the types.py module.
+defined in the models.py module.
 """
 
 # Standard
@@ -18,7 +18,7 @@ from pydantic import ValidationError
 import pytest
 
 # First-Party
-from mcpgateway.types import (
+from mcpgateway.models import (
     ClientCapabilities,
     CreateMessageResult,
     ImageContent,
